requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-experimental
langgraph
langchain-cli
langsmith
langchain-openai
streamlit

langserve
fastapi
uvicorn


sse_starlette


bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox

numpy




-e .